+++ data-browser/src/components/forms/ResourceForm.tsx
@@ -7,7 +7,7 @@ import { openURL } from '../../helpers/navigation';
 import { Button } from '../Button';
 import ResourceField from './ResourceField';
 import { ErrMessage } from './InputStyles';
-import { ResourceSelector } from './ResourceSelector';
+import { getCollectionURL, ResourceSelector } from './ResourceSelector';
 import styled from 'styled-components';
 import Field from './Field';
 import { useSettings } from '../../helpers/AppSettings';